chore: move test files in `__tests__` dir
